Pulling and formatting build logs.
