+++ .gitignore
@@ -8,6 +8,9 @@
 # Test binary, build with `go test -c`
 *.test
 
+# vscode
+.vscode
+
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 bazel-bin
